[no ticket] Fix wasm being imported on UNIFFI builds (#385)
## 🎟️ Tracking

https://bitwarden.slack.com/archives/C054ZQSBS49/p1755188403409309

## 📔 Objective

It seems https://github.com/bitwarden/sdk-internal/pull/279 accidentally
required wasm in all uniffi builds via `bitwarden-vault` ->
`bitwarden-collections` -> `bitwarden-core`. Since it is only required
on core, this leads to build issues on anything where an impl is tagged
with wasm-bindgen, but uses a struct from a third package - such as
`bitwarden-crypto` that does not have the wasm feature enabled.

Attempts to fix the build errors of:
https://github.com/bitwarden/sdk-internal/pull/383

Note: This flags out a function from the collections client. This has to
be fixed by the owning team and updated, this PR just unblocks builds
and fixes the imports.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Oscar Hinton <Hinton@users.noreply.github.com>