[PM-24127] Implement password-protected key envelope (#335)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-24127
## 📔 Objective

The current masterkey logic is complex to understand for using teams
(auth), and also prone to error. When any setting changes / gets out of
sync, such as the email, or kdf, then decryption fails. The masterkey is
further too widely scoped, used both in an authentication protocol, and
in unlock decryption.

This PR introduces a PasswordProtectedKeyEnvelope. The goal is to
protect a symmetric key with a password securely. Internally, this uses
a KDF, and the KDF settings (argon2 parameters, and random salt) are
stored on the serialized object.

That means that the only thing needed to unlock this structure is the
correct password, everything else is stored on the object, making this
process much less error prone. At the same time the interface is easier
to use.

An example is provided to show usage.

A follow-up PR will add an unlock method / enrollment for PIN based on
this new cryptographic API.

Note: Only argon2 is supported here. The PasswordProtectedKeyEnvelope's
settings are completely decoupled from the account settings, and we
don't need to provide backwards compatibility to non-recommended legacy
cryptographic algorithms (pbkdf2).

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Matt Gibson <mgibson@bitwarden.com>
Co-authored-by: Oscar Hinton <Hinton@users.noreply.github.com>
Co-authored-by: Thomas Avery <43214426+Thomas-Avery@users.noreply.github.com>