UniFFI don't panic on failed lift (#373)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

## 📔 Objective

UniFFI uses the `custom_type!` macro to convert request/response types.
The request conversion is allowed to fail which should produce an error
in the apps, but there are cases where it would produce a panic instead.
The cases are:
- If the FFI function doesn't return a result, then UniFFI produces a
panic.
- If the Error type of the function is different than the Error type in
`custom_type!`, then UniFFI produces a panic.

One way to fix this is to have the `custom_type!` calls return
`bitwarden_uniffi::error::BitwardenError`, but this isn't possible as
core and crypto can't depend on uniffi due to circular dependency
issues.

This PR makes the uniffi crate register an error conversion function
during client initialization, thanks to the use of a new `uniffi-error`
crate. It also updates all the calls to `custom_type!` to use this new
error conversion.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Thomas Avery <43214426+Thomas-Avery@users.noreply.github.com>