[PM-26534] Remove non-generic wrapper for PasswordProtectedKeyEnvelope (#488)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-26534

## 📔 Objective

The initial implementation assumed that the struct of of the
PasswordProtectedKeyEnvelope in crypto had to be generic, because the
key context is generic w.r.t. the keys used, and the keys are only then
provided by core. Because uniffy / wasm-bindgen can't handle generics, I
added a non-generic wrapper struct in core, leaving us with two
different "PasswordProtectedKeyEnvelopes".

This understanding was not correct. You don't actually need to make the
struct generic, having just two specific functions on the impl generic
to the key id's suffices, allowing us to drop the "non generic wrapper"
entirely, vastly simplifing our code here. The changes are mostly
removing code and moving the uniffi mapping to the crypto crate.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes