[PM-25012] Add data field to Cipher type, and add functions for client-side parsing of the field. (#517)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-25012

## 📔 Objective

This PR adds the `data` field to the Cipher struct, and adds a function
for parsing that field into the appropriate CipherType (Identity,
SshKey, Login, SecureNote, Card) on the client-side, rather than
consuming the server-provided values.

The function sits unused - after #455 is merged, we should consider
adding this to the `CiphersClient::get` and `CipherClient::list`
functions behind a new feature flag.