[PM-28026] Lint when holding KeyStoreContext across await points (#551)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-28026

## 📔 Objective

Clippy by default lints when trying to hold an RwLock/Mutex guard across
an await point, but that doesn't apply to custom types by default. This
PR configures clippy to also apply that lint to our own types that wrap
lock guards, which at the moment is only `KeyStoreContext`.

We only have two instances that needed fixing and both were in tests,
but some of the SM code I wrote already had to deal with it by creating
short lived contexts, so I assume it might be a problem in the future as
well:

https://github.com/bitwarden/sdk-internal/blob/main/bitwarden_license/bitwarden-sm/src/secrets/create.rs#L41-L64

The lint docs:

https://rust-lang.github.io/rust-clippy/master/index.html#await_holding_invalid_type

This is what the lint looks like:
<img width="1011" height="262" alt="image"
src="https://github.com/user-attachments/assets/77f9eaa2-f46b-460a-81b7-627e440e3f7a"
/>


## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes