[PM-27230] Introduce Account Cryptographic State (#563)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-27230
https://bitwarden.atlassian.net/browse/PM-27313

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

Client PR, fixing breaking changes:
https://github.com/bitwarden/clients/pull/17488

## 📔 Objective

Account cryptographic state describes the core cryptographic objects
making up a user. For a V1 encryption user these are: The RSA Keypair.
For a V2 user these are:
- The Public-key Encryption Keypair
- The Signature Keypair
- The Signed public key (used to bind a public-key encryption key pair
to signing keypair)
- The Signed security state (used for a safe upgrade path that prevents
insecure features)

Note: The wrapped version omits the public key and verifying key for the
public-key encryption key-pair and signature key-pair, and only includes
the wrapped private / signing keys.

Not included are:
- Keys shared to the user via an organization memebership or emergency
access
- Cipher keys
which are not part of the user's cryptography, but do interact with it.

Provided is a function to generate such a cryptographic state for v2
users, and conversion to API request models. Further, this changes SDK
initialization to be based on the account cryptographic state.

## 🚨 Breaking Changes

Please note that the public API for initializing the user's account
cryptography is updated to instead take an enum. The variants for the
enum contain the private key for V1 users, and the private key, signing
key, signed public key, signed security state for V2 users.

Aside from re-packaging into an enum variant, no other changes should be
needed on the consuming side.

This both helps prevent inconsistent states from being passed in by
enforcing consistent state via type safetey, but also cleans up the
primitive obsession anti-pattern
(https://contributing.bitwarden.com/architecture/server/#avoid-primitive-obsession)
that the crypto initialization was facing and will make future changes
much easier.

Note that since so far clients did not store the signed public key, it
is optional for now, but will be made mandatory later after clients save
it for a sufficient time.

To migrate, simply repack the existing values into the corresponding
enum. Note that previously the signed public key was not passed in, now
it is passed in (optional for now, but required later on). Please ensure
it is saved to state on sync / login.

<!-- Does this PR introduce any breaking changes? If so, please describe
the impact and migration path for clients.

If you're unsure, the automated TypeScript compatibility check will run
when you open/update this PR and provide feedback.

For breaking changes:
1. Describe what changed in the client interface
2. Explain why the change was necessary
3. Provide migration steps for client developers
4. Link to any paired client PRs if needed

Otherwise, you can remove this section. -->

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Daniel García <dani-garcia@users.noreply.github.com>
Co-authored-by: Oscar Hinton <Hinton@users.noreply.github.com>